Moved all algos into own file
