+++ include/pf/algorithms/gbfs.h
@@ -1,0 +1,13 @@
+#pragma once
+#include "../Node.h"
+
+namespace ai
+{
+  namespace path
+  {
+    std::vector<Vector2> Greedy_BFS(std::vector<NodePtr> nodes,
+                                    NodePtr              start_node,
+                                    NodePtr              end_node,
+                                    Obstacle             layer);
+  } // namespace path
+} // namespace ai
